pyproject.toml
[project]
name = "kelly-watcher"
version = "0.1.0"
description = "Polymarket intelligent copy-trading bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10",
    "httpx>=0.27",
    "joblib>=1.3",
    "numpy>=1.26",
    "pandas>=2.2",
    "py-clob-client>=0.16",
    "python-dotenv>=1.0",
    "python-telegram-bot>=21.0",
    "requests>=2.31",
    "scikit-learn>=1.4",
    "web3>=6.15",
    "xgboost>=2.0",
]

[project.scripts]
kelly-watcher = "kelly_watcher.cli:main"
main = "kelly_watcher.cli:main"
polymarket-setup = "kelly_watcher.tools.polymarket_setup:main"
rank-copytrade-wallets = "kelly_watcher.tools.rank_copytrade_wallets:main"
resolve-wallet = "kelly_watcher.tools.resolve_wallet:main"
shadow-reset = "kelly_watcher.shadow_reset:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/kelly_watcher"]
